django_backend/requirements.txt
Django==4.2.3
psycopg[binary]==3.2.9
django-redis==6.0.0
celery[redis]
djangorestframework